feat(config): add queue_io_wait_timeout_ratio, target_rate_hz, double period
Config changes (all 3 roles):
- target_period_ms: int → double (sub-ms periods supported)
- target_rate_hz: double (alternative to period, mutually exclusive)
- queue_io_wait_timeout_ratio: double, 0.1–0.5 (default 0.1)
  Controls inner retry acquire timeout as fraction of period.
  Unified formula: max(period * ratio, 10μs), no branching on policy.
- slot_acquire_timeout_ms: deprecated, LOGGER_WARN if set
- resolve_period_us() validates mutual exclusion + minimum period

Config parsing (all 3 config.cpp):
- Parse target_rate_hz, queue_io_wait_timeout_ratio
- Validate ratio range [0.1, 0.5]
- Use resolve_period_us for period/rate unification
- Deprecation warning for slot_acquire_timeout_ms

Role hosts:
- All 3 unified hosts now read config_.queue_io_wait_timeout_ratio
  (was hardcoded kDefaultQueueIoWaitRatio)
- Old hosts: static_cast<int>(target_period_ms) for chrono compatibility

Tests:
- Validation_BadTimeoutThrows → Validation_QueueIoWaitRatioOutOfRange
  (tests ratio < 0.1 and > 0.5)

1220/1220 tests pass.

+++ src/producer/producer_script_host.cpp
@@ -185,7 +185,7 @@ bool ProducerScriptHost::start_role()
     if (config_.target_period_ms > 0)
     {
         opts.loop_policy = hub::LoopPolicy::FixedRate;
-        opts.period_ms   = std::chrono::milliseconds{config_.target_period_ms};
+        opts.period_ms   = std::chrono::milliseconds{static_cast<int>(config_.target_period_ms)};
     }
     opts.ctrl_queue_max_depth = config_.ctrl_queue_max_depth;
     opts.peer_dead_timeout_ms = config_.peer_dead_timeout_ms;
@@ -676,10 +676,10 @@ void ProducerScriptHost::run_loop_()
     // When target_period_ms == 0, free-run mode: no sleep, no deadline.
     // The first iteration fires immediately; subsequent iterations sleep to hit the deadline.
     auto next_deadline = std::chrono::steady_clock::now() +
-                         std::chrono::milliseconds{config_.target_period_ms};
+                         std::chrono::milliseconds{static_cast<int>(config_.target_period_ms)};
 
     const bool is_fixed_rate = (config_.loop_timing != LoopTimingPolicy::MaxRate);
-    const auto period        = std::chrono::milliseconds{config_.target_period_ms};
+    const auto period        = std::chrono::milliseconds{static_cast<int>(config_.target_period_ms)};
 
     while (core_.running_threads.load() && !core_.shutdown_requested.load() &&
            !api_.critical_error())
